+++ src/motion_control/whole_body_control/whole_body_controller/include/whole_body_controller/prioritized_tasks.hpp
@@ -166,7 +166,7 @@ private:
     std::vector<int> tasks_vector;
 
     /// @brief
-    ContactConstraintType contact_constraint_type = ContactConstraintType::soft_kv;
+    ContactConstraintType contact_constraint_type = ContactConstraintType::rigid;
 };
 
 } // namespace wbc