+++ CHANGELOG.rst
@@ -56,9 +56,10 @@
 - The `terrain_estimator` package gives a very rough estimate of the contact plane. The local terrain height is used by the trotting planner to adjust the commanded height.
 - Important bug fix in the whole-body controller. Previously, it did not work when the orientation was very different from zero roll, pitch, and yaw.
 
-Upcoming (2023-04-02)
+Upcoming (2023-04-26)
 ------------------
 - Started using clang-tidy. The docker image installs ament-clang-tidy if the image is built with the `-d` option.
 - The estimated robot pose is displayed in RViz.
 - The state estimator works correctly and does not diverge. The filter gains are not set. The sensor noises are zero.
 - The robot can be spawned in a terrain created from an heightmap.
+- The robot measures the penetration with the terrain and increase the step height accordingly.